tests/wpt/web-platform-tests/html/browsers/offline/application-cache-api/api_swapcache_error.html

<!DOCTYPE HTML>
<html manifest="../resources/manifest/clock.manifest">
  <head>
    <title>Offline Application Cache - API_swapCache_error</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>

    <script>
      var cache = window.applicationCache;

      test(function() {
        assert_throws("INVALID_STATE_ERR", function(){cache.swapCache();});
      }, "INVALID_STATE_ERR error test")
    </script>
  </body>
</html>
